update alembic script
